requirements.txt
boto3==1.33.5
botocore==1.33.5
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.1.1
cryptography==41.0.7
cycler==0.12.1
docutils==0.20.1
fonttools==4.45.1
idna==3.6
importlib-metadata==6.8.0
importlib-resources==6.1.1
jaraco.classes==3.3.0
jeepney==0.8.0
jmespath==1.0.1
keyring==24.3.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
mathplot==0.1
matplot==0.1.9
matplotlib==3.7.4
mdurl==0.1.2
more-itertools==10.1.0
nh3==0.2.14
numpy==1.24.4
opencv-python-headless==4.8.1.78
packaging==23.2
Pillow==10.1.0
pkginfo==1.9.6
pycparser==2.21
Pygments==2.17.2
pyloco==0.0.139
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
readme-renderer==42.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.0
s3transfer==0.8.2
SecretStorage==3.3.3
SimpleWebSocketServer==0.1.2
six==1.16.0
twine==4.0.2
typing==3.7.4.3
typing_extensions==4.8.0
urllib3==1.26.18
ushlex==0.99.1
websocket-client==1.6.4
zipp==3.17.0
